requirements.txt
beautifulsoup4==4.6.3
certifi==2018.8.13
chardet==3.0.4
Django==2.1
fake-useragent==0.1.10
future==0.16.0
Google-Search-API==1.1.14
idna==2.7
multidict==4.4.0a38
pytz==2018.5
PyYAML==4.2b4
requests==2.19.1
selenium==2.53.6
six==1.11.0
Unidecode==1.0.22
urllib3==1.23
vcrpy==1.13.0
wrapt==1.10.11
yarl==1.2.6
